+++ dart/lib/src/sentry_user_feedback.dart
@@ -3,6 +3,7 @@ import 'package:meta/meta.dart';
 import 'protocol.dart';
 import 'protocol/access_aware_map.dart';
 
+@Deprecated('Will be removed in a future version. Use [SentryFeedback] instead')
 class SentryUserFeedback {
   SentryUserFeedback({
     required this.eventId,