+++ drift/test/sentry_database_test.dart
@@ -1,4 +1,4 @@
-// ignore_for_file: invalid_use_of_internal_member
+// ignore_for_file: invalid_use_of_internal_member, library_annotations
 
 @TestOn('vm')