+++ flutter/android/src/test/kotlin/io/sentry/flutter/SentryFlutterTest.kt
@@ -69,7 +69,7 @@ class SentryFlutterTest {
     assertEquals("0000", fixture.options.proxy?.pass)
 
     assertEquals(0.5, fixture.options.experimental.sessionReplay.sessionSampleRate)
-    assertEquals(0.6, fixture.options.experimental.sessionReplay.errorSampleRate)
+    assertEquals(0.6, fixture.options.experimental.sessionReplay.onErrorSampleRate)
 
     // Note: these are currently read-only in SentryReplayOptions so we're only asserting the default values here to
     // know when there's a change in the native SDK, as it may require a manual change in the Flutter implementation.