+++ flutter/lib/src/native/sentry_native_channel.dart
@@ -87,11 +87,17 @@ class SentryNativeChannel
   }
 
   @override
+  bool get supportsCaptureEnvelope => true;
+
+  @override
   Future<void> captureEnvelope(
       Uint8List envelopeData, bool containsUnhandledException) {
     return channel.invokeMethod(
         'captureEnvelope', [envelopeData, containsUnhandledException]);
   }
+
+  @override
+  bool get supportsLoadContexts => true;
 
   @override
   Future<Map<String, dynamic>?> loadContexts() =>
@@ -178,11 +184,17 @@ class SentryNativeChannel
       });
 
   @override
-  Future<List<DebugImage>?> loadDebugImages(Set<String> instructionAddresses) =>
+  Future<List<DebugImage>?> loadDebugImages(SentryStackTrace stackTrace) =>
       tryCatchAsync('loadDebugImages', () async {
-        final images =
-            await channel.invokeListMethod<Map<dynamic, dynamic>>(
-                'loadImageList', instructionAddresses);
+        Set<String> instructionAddresses = {};
+        for (var frame in stackTrace.frames) {
+          if (frame.instructionAddr != null) {
+            instructionAddresses.add(frame.instructionAddr!);
+          }
+        }
+
+        final images = await channel.invokeListMethod<Map<dynamic, dynamic>>(
+            'loadImageList', instructionAddresses.toList());
         return images
             ?.map((e) => e.cast<String, dynamic>())
             .map(DebugImage.fromJson)